move into list items
