+++ mobile/src/components/common/UsernameWithAvatar.js
@@ -1,14 +1,13 @@
 import { PropTypes } from 'prop-types';
 import React from 'react';
-import { Pressable, StyleSheet, View } from 'react-native';
+import { StyleSheet, View } from 'react-native';
 import Avatar from './Avatar';
 import { Title } from './typography';
 
 const _style = StyleSheet.create({
   usernameWithAvatar: {
     display: 'flex',
-    flexDirection: 'row',
-    paddingHorizontal: 10
+    flexDirection: 'row'
   },
   avatar: {
     marginRight: 15
@@ -18,14 +17,12 @@ const _style = StyleSheet.create({
   }
 });
 
-const UsernameWithAvatar = ({ user, size, onPress }) => {
+const UsernameWithAvatar = ({ user, size }) => {
   return (
-    <Pressable onPress={() => onPress()}>
-      <View style={_style.usernameWithAvatar}>
-        <Avatar style={_style.avatar} size={size} user={user} />
-        <Title style={_style.username}>{user.username}</Title>
-      </View>
-    </Pressable>
+    <View style={_style.usernameWithAvatar}>
+      <Avatar style={_style.avatar} size={size} user={user} />
+      <Title style={_style.username}>{user.username}</Title>
+    </View>
   );
 };
