+++ mobile/src/components/common/index.js
@@ -5,13 +5,13 @@ import ErrorMessage from './ErrorMessage';
 import { FormLayout } from './forms';
 import { ScreenContainer } from './layout';
 import { CardList, Table } from './lists';
+import { TmdbListItem, UserListItem } from './lists/list_items';
 import { BottomSheet } from './menus';
 import { Poster } from './poster';
 import Select from './Select';
 import Spinner from './Spinner';
 import { Surface } from './surface';
 import TextInput from './TextInput';
-import { TmdbListItem } from './tmdb';
 import { Headline, Text, Title } from './typography';
 import UserCard from './UserCard';
 import UsernameWithAvatar from './UsernameWithAvatar';
@@ -26,13 +26,14 @@ export {
   ScreenContainer,
   CardList,
   Table,
+  TmdbListItem,
+  UserListItem,
   BottomSheet,
   Poster,
   Select,
   Spinner,
   Surface,
   TextInput,
-  TmdbListItem,
   Headline,
   Text,
   Title,
