+++ mobile/src/components/common/lists/list_items/tmdb/TmdbListItem.js
@@ -1,8 +1,9 @@
 import { Poster } from 'components/common/poster';
 import React from 'react';
-import { Pressable, StyleSheet, View } from 'react-native';
+import { StyleSheet, View } from 'react-native';
+import { TouchableHighlight } from 'react-native-gesture-handler';
 import { navigate } from 'shared/RootNavigation';
-import { Title } from '../typography';
+import { Title } from '../../../typography';
 import MovieInfo from './MovieInfo';
 import PersonInfo from './PersonInfo';
 
@@ -10,7 +11,7 @@ const _style = StyleSheet.create({
   resultItem: {
     display: 'flex',
     flexDirection: 'row',
-    paddingHorizontal: 10
+    padding: 10
   },
   info: {
     flexGrow: 1,
@@ -25,7 +26,7 @@ const _style = StyleSheet.create({
 
 const TmdbListItem = ({ header, imageLink, item, type }) => {
   return (
-    <Pressable onPress={() => type === 'movie' && navigate('MovieDetails', { movieId: item.id, title: item.title })}>
+    <TouchableHighlight onPress={() => type === 'movie' && navigate('MovieDetails', { movieId: item.id, title: item.title })}>
       <View style={_style.resultItem}>
         <Poster size='sm' title={header} source={{ uri: imageLink }}/>
         <View style={_style.info}>
@@ -37,7 +38,7 @@ const TmdbListItem = ({ header, imageLink, item, type }) => {
           </View>
         </View>
       </View>
-    </Pressable>
+    </TouchableHighlight>
   );
 };
