+++ mobile/src/components/common/lists/list_items/tmdb/index.js
