+++ mobile/src/components/common/lists/table/Table.js
@@ -1,5 +1,4 @@
 import Divider from 'components/common/Divider';
-import _ from 'lodash';
 import React from 'react';
 import { StyleSheet, View } from 'react-native';
 import { FlatList } from 'react-native-gesture-handler';
@@ -11,33 +10,11 @@ const _style = StyleSheet.create({
     overflow: 'visible',
     height: '100%'
   },
-  tableRow: {
-    marginVertical: 10
-  },
-  tableRowFirst: {
-    marginTop: 0,
-    marginBottom: 10
-  },
-  tableRowLast: {
-    marginTop: 10,
-    marginBottom: 0
-  },
   emptyState: {
     marginTop: 10,
     marginHorizontal: 10
   }
 });
-
-const getActiveRowStyle = (index, size) => {
-  switch (index) {
-  case 0:
-    return _style.tableRowFirst;
-  case (size - 1):
-    return _style.tableRowLast;
-  default:
-    return _style.tableRow;
-  }
-};
 
 const Table = ({ models, columns, Header, localization }) => {
   return (
@@ -46,11 +23,8 @@ const Table = ({ models, columns, Header, localization }) => {
         style={_style.flatList}
         data={models}
         renderItem={({ item, index }) => {
-          const activeRowStyle = getActiveRowStyle(index, _.size(models));
           return (
-            <View key={index}>
-              <TableRow model={item} columns={columns} style={activeRowStyle}/>
-            </View>
+            <TableRow key={index} model={item} columns={columns}/>
           );
         }}
         ItemSeparatorComponent={Divider}
@@ -60,7 +34,6 @@ const Table = ({ models, columns, Header, localization }) => {
           </View>
         )}
         ListHeaderComponent={Header}
-        ListFooterComponent={() => <View style={{ marginBottom: 10 }}/>}
       />
     </View>
   );
